Move sharable integration tests boiler plate to utils
